optimize

+++ src/query/service/src/interpreters/interpreter_table_drop_column.rs
@@ -27,7 +27,6 @@ use databend_storages_common_table_meta::table::OPT_KEY_BLOOM_INDEX_COLUMNS;
 
 use crate::interpreters::common::check_referenced_computed_columns;
 use crate::interpreters::interpreter_table_add_column::commit_table_meta;
-use crate::interpreters::util::check_column_has_policy;
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
 use crate::sessions::QueryContext;
@@ -87,12 +86,12 @@ impl Interpreter for DropTableColumnInterpreter {
         let table_schema = table_info.schema();
         let field = table_schema.field_with_name(self.plan.column.as_str())?;
 
-        check_column_has_policy(&table_info.meta, &field.column_id).map_err(|_| {
-            ErrorCode::AlterTableError(format!(
+        if table_info.meta.is_column_reference_policy(&field.column_id) {
+            return Err(ErrorCode::AlterTableError(format!(
                 "Cannot drop column '{}' which is associated with a security policy",
                 self.plan.column.as_str()
-            ))
-        })?;
+            )));
+        }
         if field.computed_expr().is_none() {
             let mut schema: DataSchema = table_info.schema().into();
             schema.drop_column(self.plan.column.as_str())?;
